+++ build.rs
@@ -1,6 +1,6 @@
 fn main() {
     println!("cargo:rerun-if-env-changed=GIT_COMMIT");
     if let Ok(sha) = std::env::var("GIT_COMMIT") {
-        println!("cargo:rustc-env=GIT_COMMIT_HASH={}", sha);
+        println!("cargo:rustc-env=GIT_COMMIT_HASH={sha}");
     }
 }